Properly resolve generic struct/union instance aliases
There were issues with resolving things like Opt<*T> because aliases wouldn't be resolved recursively. Since Opt<*T> is an instance of Opt<T> where {T: *T} when trying to resolve the type Opt<*T> from Opt<*X> we would expect to see that the final result was that {T: X}, but this wasn't happening, but it now does. This required that a step was when resolving generic struct/union instances where every type in the alias map also has to be resolved. This ensures that each alias assigns generics only in the base case.