Removed the `split` keyword and op
Being able to split a struture was going to limit the safety that
the language would be able to provide. If Haystack is to be able
to provide some form of RAII, than it's imporant that a structure
is guaranteed to be destoyed correctly. If a user can destructure
a struct, than there's no way to ensure that any given structure
will be destroyed appropriately.

Furthermore it's possible to achieve a similar affect with named
variables.