Added initial implementation of destructors
Destructors are invoked at the end of scope for each type that is
on a frame. The current implementation is not fast, as it copies
each member of a struct in order to call the destructor for each
member.

Destructors are auto generated for types which do not provide a
destructor.