Throw it all away
Ok, it was becoming really cumbersome to keep working on haystack with where it was.

There's a lot of improvements I want to get done, but I'm probably better off throwing a lot of it away.
It's unfortunate, but hey, lessons can be learned from.

So what's in this change.
Let's start off by working without the parser. It'll be annoying for a little bit, but hopefully it will
be easier to port my type system prototypes over to, and make sure that the internals are really well
sorted.

Here, I've setup expressions and statements, and have an "example" hello world program to prove I can
represent code and have it compile.

Expressions and Statements have been overhauled to allow for metadata to be attached to them, which
should make it a lot easier to represent different passes. Additionally, expressions can now have custom
extensions which make it easier to represent intermediate steps separately from the base expressions.